Update the generation of configuration files to match CF config
We have previously used Stunnel in Cloudfoundry to secure access to
infra within the VPC.  This buildpack does the same thing for Heroku Redis add-ons.

While we are not using a Redis add-on (at this point in time) we can leverage
the buildpack to connect to our infra.

The changes in this file adapt the config generated by the buildpack to
match that which we used with CF.  (Note the PSK auth)