Simplify the run-stunnel script
There is a lot of code that is duplicate of that in the start-stunnel
script.  This can be removed by importing start-stunnel and changing the
bits that matter.

What's more, some of the code included in this script is specific to its
application to Redis and does not apply to our usage of stunnel.  It can be
removed.